mount-utils: Detect potential stale file handle
This is a fix for the issue #97031. It may happen that a subpath mount
point on a networking FS gets deleted on the server but in the case of
some file systems (CIFS) it would make the stat syscall end with ENOENT,
preventing the mount point to be cleaned up and pods using the mount
can't be cleanly deleted.

This situation can be detected by calling also access syscall on the
same path: if there is a discrepancy between stat and access results,
treating the mount as corrupted allows for successful cleanup and
unmount.
